requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
openai==0.27.10
python-dotenv==1.0.0
groq
selenium==4.27.1
httpx<=0.27.2
